Move the new InnerJoinOperation
