+++ java-vtl-script/src/main/java/no/ssb/vtl/script/support/VTLPrintStream.java
@@ -9,9 +9,9 @@ package no.ssb.vtl.script.support;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,7 +20,12 @@ package no.ssb.vtl.script.support;
  * =========================LICENSE_END==================================
  */
 
-import no.ssb.vtl.model.*;
+import no.ssb.vtl.model.Component;
+import no.ssb.vtl.model.DataPoint;
+import no.ssb.vtl.model.DataStructure;
+import no.ssb.vtl.model.Dataset;
+import no.ssb.vtl.model.Order;
+import no.ssb.vtl.model.VTLObject;
 
 import java.io.OutputStream;
 import java.io.PrintStream;
